feat(kerberoast): move get_conditional_action to helper and implement specific user kerberoasting
